Extract core logic of AddAssemblyReferences to ScriptExtensionsProjectMetadata
